handle changes to protean

+++ src/protean_sqlalchemy/cli.py
@@ -16,7 +16,8 @@ Why does this file exist, and why not put this in __main__?
 """
 import click
 
-from protean.core.repository import repo_factory
+from protean.conf import active_config
+from protean.core.repository import repo
 
 from protean_sqlalchemy.repository import SqlalchemySchema
 
@@ -33,8 +34,10 @@ def create_tables():
     click.echo('Creating all tables for registered entities')
 
     # Create all the tables
-    for conn in repo_factory._connections.values():
-        SqlalchemySchema.metadata.create_all(conn.bind)
+    for conn_name, conn in repo.connections.items():
+        if active_config['REPOSITORIES'][
+                conn_name]['PROVIDER'] == 'protean_sqlalchemy.repository':
+            SqlalchemySchema.metadata.create_all(conn.bind)
 
 
 @main.command()
@@ -42,6 +45,8 @@ def drop_tables():
     """ Command to drop all tables for registered entities"""
     click.echo('Dropping all tables for registered entities')
 
-    # Create all the tables
-    for conn in repo_factory._connections.values():
-        SqlalchemySchema.metadata.drop_all(conn.bind)
+    # Delete all the tables
+    for conn_name, conn in repo.connections.items():
+        if active_config['REPOSITORIES'][
+                conn_name]['PROVIDER'] == 'protean_sqlalchemy.repository':
+            SqlalchemySchema.metadata.drop_all(conn.bind)
